+++ packages/electron/src/renderer/components/AgentMode/AgentMode.tsx
@@ -43,7 +43,7 @@ import {
   isRestoringNavigationAtom,
 } from '../../store';
 import { errorNotificationService } from '../../services/ErrorNotificationService';
-import { initWorkstreamState, loadWorkstreamStates, workstreamStateAtom, workstreamActiveChildAtom, setWorktreeActiveSessionAtom } from '../../store/atoms/workstreamState';
+import { initWorkstreamState, loadWorkstreamStates, workstreamStateAtom, workstreamActiveChildAtom, setWorkstreamActiveChildAtom, setWorktreeActiveSessionAtom } from '../../store/atoms/workstreamState';
 import { initSessionStateListeners } from '../../store/sessionStateListeners';
 
 export interface AgentModeRef {